Drop hard dependency on Acquisition and speed up removeAllProxies
Acquisition is still used if available, and is still a test
dependency, but it is not a hard installation dependency anymore.

The function proxy.removeAllProxies is called typically at least twice
for every object we externalize (when getting oids), making its speed
important. The function was naievly written and quite slow.

For a deeply wrapped object (10 layers of proxies), the new
implementation is roughly 2.9x faster (16.9s vs 6.0s). For (the common
case of) an object with no wrapper, it's about 2.2x faster (2.23s vs
1.05s). (All numbers in CPython 2.7.)