More typing in datastructures.py
This nets us a 4% improvement or so in externalization, maybe a tiny
slowdown in internalization (but I expect that to be fixed when I type
that module).

We could do better, but the class attribute _ext_excluded_out_ and the
like are slowing us down, and we have some code that doesn't keep
their usual types (mostly in housing). A metaclass may be able to fix
that, because they all seem to be defined at the class level.