Learn how to do enumerated constants as class bodies in cython
Define them as an instance, make an accessor function, type the result
of the accessor in the pxd. This avoids needing to get the type in the
python file, which is what causes the leading underscore problems.

This avoids the need for re-enumerating all the constants while still
being direct access (yay!). Speeds things up by another percent or two
in datastructures.py.