Fix the interface resolution order.
Make ILocatedExternalDict/List extend the more modern interfaces.

Make the implementation classes keep providing the legacy interfaces to the extent possible (everything for Dict, only some for List).

In practice I don't expect this to make a difference to client code; I think registrations for the legacy interfaces that were expected to be hit for instances of ILocatedExternal* would be rare to non-existant.

Fixes #105